pyproject.toml
[tool.poetry]
name = "otakustan"
version = "0.1.0"
description = ""
authors = ["repl.it user <replituser@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
"discord.py" = "^1.5.0"
python-decouple = "^3.3"
flask = "^1.1.2"
requests = "^2.24.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
